feat(elements): Move some files around, create initial package structure
